requirements.txt
appier
appnope
black
coveralls
django-bootstrap-icons
django-debug-toolbar
djangorestframework
google
google-genai
pipdeptree
PyMuPDF
pytest-cov
pytest-django
python-dotenv